Add "-mhwacha4" to GCC
All this does is set __riscv_hwacha4 to be defined in user code.  As a
side-effect, it also defines TARGET_HWACHA4, which allows GCC to
control codegen for Hwacha v4 (but GCC doesn't do any codegen, so that
doesn't do anything).  I did this by pattern matching some existing
stuff, but it looks pretty much right (and at generates the define iff
passed the argument).

This allows me to ifdef in vector assembly for RISC-V benchmarks, so I
can mostly run the same code on RISC-V and Intel.